파일 구조 변경 및 UI 요소 제거
